New syntax client.connect() code got closed up in the client module.
